Relaxes validation for `macaddr` and `macaddr8`
- Opting to let PostgreSQL reject invalid lengths.
- Adds a test to verify which error is thrown from the backend.
