+++ src/Npgsql/TypeHandlers/NetworkHandlers/MacaddrHandler.cs
@@ -1,4 +1,5 @@
 ﻿#region License
+
 // The PostgreSQL License
 //
 // Copyright (C) 2018 The Npgsql Development Team
@@ -19,12 +20,11 @@
 // AND FITNESS FOR A PARTICULAR PURPOSE. THE SOFTWARE PROVIDED HEREUNDER IS
 // ON AN "AS IS" BASIS, AND THE NPGSQL DEVELOPMENT TEAM HAS NO OBLIGATIONS
 // TO PROVIDE MAINTENANCE, SUPPORT, UPDATES, ENHANCEMENTS, OR MODIFICATIONS.
+
 #endregion
 
-using System;
 using System.Diagnostics;
 using System.Net.NetworkInformation;
-using JetBrains.Annotations;
 using Npgsql.BackendMessages;
 using Npgsql.TypeHandling;
 using Npgsql.TypeMapping;
@@ -32,21 +32,26 @@ using NpgsqlTypes;
 
 namespace Npgsql.TypeHandlers.NetworkHandlers
 {
+    /// <summary>
+    /// Defines the type handler for macaddr and macaddr8.
+    /// </summary>
     /// <remarks>
     /// http://www.postgresql.org/docs/current/static/datatype-net-types.html
     /// </remarks>
+    [TypeMapping("macaddr8", NpgsqlDbType.MacAddr8)]
     [TypeMapping("macaddr", NpgsqlDbType.MacAddr, typeof(PhysicalAddress))]
     class MacaddrHandler : NpgsqlSimpleTypeHandler<PhysicalAddress>
     {
         #region Read
 
+        /// <inheritdoc />
         public override PhysicalAddress Read(NpgsqlReadBuffer buf, int len, FieldDescription fieldDescription = null)
         {
-            Debug.Assert(len == 6);
+            Debug.Assert(len == 6 || len == 8);
 
-            var bytes = new byte[6];
+            var bytes = new byte[len];
 
-            buf.ReadBytes(bytes, 0, 6);
+            buf.ReadBytes(bytes, 0, len);
             return new PhysicalAddress(bytes);
         }
 
@@ -54,13 +59,16 @@ namespace Npgsql.TypeHandlers.NetworkHandlers
 
         #region Write
 
+        /// <inheritdoc />
         public override int ValidateAndGetLength(PhysicalAddress value, NpgsqlParameter parameter)
-             => value.GetAddressBytes().Length == 6
-                 ? 6
-                 : throw new FormatException("MAC addresses must have length 6 in PostgreSQL");
+            => value.GetAddressBytes().Length;
 
+        /// <inheritdoc />
         public override void Write(PhysicalAddress value, NpgsqlWriteBuffer buf, NpgsqlParameter parameter)
-            => buf.WriteBytes(value.GetAddressBytes(), 0, 6);
+        {
+            var bytes = value.GetAddressBytes();
+            buf.WriteBytes(bytes, 0, bytes.Length);
+        }
 
         #endregion Write
     }
